bug(fix alembic files): Update alembic files for room model
 - Delete errored alembic file

 - Generate new alembic file

[Fixes #160193943]
