+++ alembic/versions/5195102fd25c_remove_unique_attribute_from_room_name.py
@@ -1,0 +1,28 @@
+"""Remove unique attribute from room name
+
+Revision ID: 5195102fd25c
+Revises: 4dc677f14096
+Create Date: 2018-09-12 13:32:25.438486
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '5195102fd25c'
+down_revision = '4dc677f14096'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('rooms_name_key', 'rooms', type_='unique')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_unique_constraint('rooms_name_key', 'rooms', ['name'])
+    # ### end Alembic commands ###
